[Learners] Implemented update of learning rates alphas by state-ACTION visit count, not only by STATE visit count
Goal: Have a more suitable update strength of the Q(s,a) value function, e.g. so that a larger update strength is used when a new action has been used for a given state that was visited many times already but with other actions taken.

Summary: There is now a new array containing the learning rates alpha for each possible state and action, which are updated based on the respective state-action visit count.
In both MC and TD learners, calls to the visit counts update have been updated to the new tracking mechanism.

This is a partial commit, so correct execution is not guaranteed.

DETAILS:
- learners/episodic/discrete/__init__.py: Learner class:
  - New attributes to handle the learning rates _alphas2 defined for each state and action. All the attributes that keep track of visit counts to adjust the learning rate start with prefix "_action" (e.g. _action_counts_over_all_episodes).
  - The method update_state_counts() was renamed to update_visit_counts() because both the state- and state-action visit counts are updated. A new argument is now received, namely the current taken action.