[Learners] Implemented agent's discovery of environment states and storage of EXIT states as activation set in the FV learner
Goal: Avoid cheating in certain learning situations where the total number of states was used in a criterion (e.g. deciding when to stop growing the absorption set A in the FV learner) or when the true activation set was used to choose the particle start states for the FV simulation, even though NO exit states were observed by the agent during the initial exploration of the environment. Now the activation_set attribute of the FV learner stores the set of OBSERVED exit states from A.

Summary: The logic for storing backup EXIT states is the following: they are stored as the activation set of the FV learner as keys of a dictionary that also stores their probability computed as the proportion of times they were observed as EXIT states from A during the initial exploration of the environment. When NO exit states are observed in a future exploration of the environment (i.e. on a different policy learning step), the set of EXIT states is defined as the subset of the activation set that is NOT part of the absorption set. If this subset is NOT empty, the particle start states are selected using a UNIFORM distribution as selection probability. If the subset is empty, the FV simulation is NOT run, since the learner doesn't know where the FV particles should start. Note that the selection probability distribution is chosen as UNIFORM so that the start states are distributed homogeneously among the backup EXIT states which could help BALANCE a very skewed distribution from a previous excursion that may bias the concentration of particles in regions where they could get stuck.

This is a partial commit, so correct execution is not guaranteed.

DETAILS:
- agents/learners/__init__.py: GenericLearner class:
  - New attribute environment_set that stores the set of known states by the agent.
  - New method updateKnownEnvironmentSet() that updatese the set of known states by the agent.
  - New method getKnownEnvironmentSet() that retrieves the set of known states by the agent.

- discrete/fv.py:
  - setAbsorptionSet(): removed the update of the activation set which was computed from the true transtion probabilities. Reason: this information may lead to cheating in the FV learning process because we were accidentally choosing the particle start states from this set which might be unknown to the agent if no EXIT states from A were observed during the initial exploration of the environment.
  - New method setActivationSet() that stores the set of EXIT states from the absorption set A so that it can be used as particle start states in certain cases where no EXIT states are observed during the initial exploration of the environment (e.g. this serves as a backup of the exit states in those flawed situations).
  - getActivatonSet(): new parameter `retrieve_probabilities=False` which can be set to True when the activation set stored in the learner (by the new setActivationSet() method) stores the set of states as keys in a dictionary defining their probability of being observed as EXIT states from A, which is then used as selection probability for the particle start states for the FV simulation.

- simulators/discrete.py:
  - update_absorption_set_if_not_too_large() internal function of _estimate_value_functions_and_expected_reward_fv(): the condition to stop updating (growing) the absorption set A is now based on a percentage of the number of KNOWN states as opposed to the number of states in the environment, as the agent is NOT supposed to know the number of states in the environment.

  - _run_simulation_fv(): the parsing of the start states distribution is now done only ONCE instead of as many times as number of particles in the FV system.