[Gridworlds] Fixed the dynamics of the gridworld with wind
The implementation of the transition probability matrix of the gridworld with wind was incorrect because it allowed for some leakage, i.e. the agent could appear in impossible NON-accessible locations (a problem I realized when running simulations on a gridworld with obstacles having non-accessible regions from the usual start state).
The problem was caused by the fact that the `dict_next_state_dictated_by_wind_for_each_direction` dictionary, defining the next state dictated by the action taken + the influence of the wind, did NOT check whether the next state *without* wind was a feasible next state, and thus the agent could end up at a state that required passing through a non-feasible state in order to be reached.
New unit tests are now in place to verify the correct behaviour of an agent in a windy gridworld: see Test_Support_EnvGridworld2D_WithObstaclesAndWind.test_move().

Note: Obstacles in gridworlds are now shown in gray, instead of black, so that it is easier to read the gridworld itself, as the grid lines now become more visible.

This is a partial commit, so correct execution is not guaranteed.