[Learners] Properly implemented TD(Lambda) learning for CONTINUING learning tasks
Summary: Until now TD(Lambda) learning on CONTINUING learning tasks was not correctly implemented because the eligibility traces were reset to zero at the end of an "episode", i.e. when the agent reached a terminal state in the original episodic problem. Now the learners were adapted so that resetting the eligibility traces (one for V and one for Q) as well as the information defining the learning rates alpha applied to each state and state-action of the eligibility traces is only done under EPISODIC learning tasks or when reset_episode=True, a new parameter added to the self._reset_at_start_of_episode() method of GenericLearner, Learner, LeaTDLambda, LeaTDLambdaAdaptive, LeaMCLambda.
Note that the parameter name is not really related to the eligibility traces concept because it also affects the reset of visit counts that define the alpha learning rates, which are reset by the superclass learners who do not know anything about eligibility traces.
Note also that this special reset for CONTINUING learning tasks is NOT performed in LeaMCLambda because the results obtained by doing so (see currently commented implementation in LeaMCLambda._reset_at_start_of_episode()) are MUCH worse than not doing such tailored reset for CONTINUNING learning tasks. This is based on plotting the evolution (plot=True in sim_mc.run()) of the value function estimate by Test_EstDifferentialValueFunctions_EnvGridworld2DWithObstacles.test_Env_PolRandomWalk_MetMC() in test_estimators_discretetime.py. See also the TODO task dated 2025/07/10 in mc.py.

Note: The effect of this change if VERY VERY SMALL, at least in the environments I normally deal with, with very sparse reward landscape. Specifically, the observation is based on the plots generated by the two tests running TD(Lambda) in test_estimators_discretetime.py: test_EnvGridworld1DOneTerminal_PolOptimal_MetTDLambda() and test_EnvGridworld1DOneTerminal_PolRandomWalk_MetTDLambda(), where we see almost no difference in the estimated state value function.

This is a partial commit, so correct execution is not guaranteed.
However, all 69 unit tests should pass.

DETAILS:
- learners/episodic/discrete/__init__.py:
  - New method _reset_trajectory() which allows to perform that particular reset independently of resetting the other pieces of information of the learner. This will be used when we propertly implement FVAC(Lambda) in the near future (see reason why it is wrong in the entry below for td.py).

  - _reset_at_start_of_episode(): New reset_episode=True which allows resetting the visit counts only when reset_episode=True or under the EPISODIC learning task.

- learners/episodic/discrete/td.py:
  - New method reset_traces() that is responsible for resetting the eligibility traces. This was separated in case we want to call just the reset of the eligibility traces but not the reset of other pieces of information that are normally reset at the start of an episode, such as the alpha learning rates or the visit counts. This in fact will be done when we propertly implement FVAC(Lambda) in the near future, since the original implementation committed on 08-Jul-2025 is actually WRONG (because it wrongly resets the visit counts when a particle is absorbed and this should not be the case, as visit counts define learning rates and they should be accounted for as an OVERALL measure of the whole particle system, because they measure how many times an update of the corresponding value functions was carried out already, thus triggering a reduction of the learning rate alpha for the corresponding state or state-action).

  - _reset_at_start_of_episode(): New reset_episode=True which allows resetting the eligibility traces only when reset_episode=True or under the EPISODIC learning task.

- learners/episodic/discrete/mc.py:
  - _reset_at_start_of_episode(): New reset_episode=True to align the method with the signature of the other _reset_at_start_of_episode() methods in learner classes. However, this parameter is currently NOT used because I haven't yet analyzed how the return values should be reset in CONTINUING learning tasks for MC(Lambda) learners.
  See also TODO task in learn_mc().

- test/test_estimators_discretetime.py:
  - Updated the expected results on the two existing TD(lambda) tests run on the 1D gridworld environment.