[FVAC] Fixed computation of E(T_A) to avoid overestimation of average reward when not enough cycles are observed
Summary: When not enough number of cycles are observed during the initial exploration of the environment used to estimate E(T_A) (the condition is tested against dict_params_simul['min_num_cycles_for_expectations'] which, when not given is set to global variable MIN_MUM_CYCLES_FOR_EXPECTATIONS), the censored observation is used as one more sample of the reabsorption cycle time.
To avoid that this threshold value is zero, parameter `min_num_cycles_for_expectations` when calling discrete.Simulator.run() was changed from 0 to `None` in tests.py.

This is a partial commit, so correct execution is not guaranteed.
However, all 69 unit tests should pass.