[FVAC] Redesigned randomness in the A-estimation simulation, the INI exploration, and management of hyperparameter T when linked to A
Summary: The following changes were carried out regarding the initial exploration of the environment related to the estimation of the absorption set A and to the initial exploration used to estimate E(T_A) with the goal of (i) increasing variability in explorations and thus reduce the probability that the FV system gets stuck on a situation where NO EXIT states are observed from A, and (ii) avoid using too many learning steps during the initial explorations so that competing benchmark methods do not get a too large budget to use when FVAC fails.
The changes are:
a) Set parameter `epsilon_random_action` of discrete.Simulator.run_exploration() to the value passed by the user (e.g. typically 0.1) when running the simulation that estimates the absorption set A. Even though we could well use epsilon_random_action=0.0 (as was the case until now), because this simulation does NOT learn value functions, therefore there is no need to explore beyond the randomness inherent in the policy, using a positive epsilon_random_action value may increase variability of the estimated A, thus possibly allowing higher chances to observe EXIT states from A.

b) When NO EXIT states are observed from A, the start state for the E(T_A) simulation is now selected at every exploration performed with increased T value, again in order to increase variability in the explorations run which could increase the chances of observing EXIT states from A.

c) At the start of EACH policy learning step, the value of T is reset to its originally set value by the user, in order to avoid having at too large budget in future policy learning steps when such large budget is possibly no longer needed because the policy changed closer to an optimal policy.

This is a partial commit, so correct execution is not guaranteed.
However, all 69 unit tests should pass.