[Execution] The #events observed during the A-estimation simulation are now considered as part of the steps run by FVAC
This increases the budget used for the benchmark method TDAC. I think this is fairer, even though during the A-estimation simulation NO learning of value functions is carried out (as opposed to TDAC, which estimates value functions at EVERY simulation step).

This is a partial commit, so correct execution is not guaranteed.
However, all 73 unit tests pass (tested on a clean repository, after stashing the non-committed changes).