[FVAC] Growth of absorption set A now limited to 90% of known states (as opposed to no check)
Goal: Avoid possible "intense" reception of reward signals during the FV simulation, which could make learning unstable... e.g. producing UNLEARNING. The effect on unlearning was actually observed, either because it did not happen at all or because it happened later than the case when no check on the growth of A is performed.

All 73 unit tests pass.