[Simulator] Reference state value function is now mean(V) as opposed to V[0], both in RMSE calculations and in plots
Goal: Use a more sensible value to evaluate the RMSE of the V(s) estimation and to plot, as using V[0] as reference value (respectively V_true[0] for the true V(s) function) always gives zero error for V[0] (as the values used for the RMSE calculation are usually `V[s] - V[0]` and `V_true[s] - V_true[0]`, which is zero for s = 0 in both estimated and true V(s).

New method _compute_function_reference_values() that now unifies the calculation of the V(s) reference values, for the estimated and the true V(s).

This is a partial commit, so correct execution is not guaranteed.

DETAILS:
- utils.py:
  - rmse(): Now all averages are computed by ignoring NaN values (e.g. np.nanmean(), np.nansum() are now used).
  Goal: Allow for missing values in the true V(s) when e.g. a gridworld has obstacles so that no error for those states contribute to the RMSE value that is used to evaluate the quality of an estimated V(s).

- test/test_utils.py:
  - New plot_estimated_state_value_function() function that was taken and adapted from its previous location in test_estimators_discretetime.py.
	The following adaptations were done:
    - Plotted value functions (estimated and true, if existing and stored in the given environment) are now corrected by their respective mean (instead of V[0]), in line with the changes performed in discrete.py in this respect.
    - The state counts and the learning rates per state can now be plotted on a secondary axis, so that we know the sample size behind the estimation of each state value V(s).