[Policy] The neural network policy parameterization now defines the torch seed only at object construction
Summary: The PolNN implementing a policy parameterized by a neural network now defines the torch random seed ONLY at the construction of the object.
Goal: Avoid repetition of the very same trajectories when the policy doesn't change, because the init_policy() method was setting the seed again every time it was called!
Note: This might solve the problem I observed in July about generating exactly the same trajectories with discrete.Simulator.run_exploration() when no EXIT states from A were observed in a labyrinth problem, making learning be stuck because the Fleming-Viot simulation was never run.

Although this is a partial commit, everything should work fine, as this change is quite isolated.