[Learners] Tabular advantage function and TRUE GAE (Generalized Advantage Estimation), for faster learning with TD(lambda)
Goal: Learn the advantage function using TD(lambda), thus learning faster than with TD(0).

Summary: The advantage function is no longer estimated using function approximation *directly*, but through the delta(V) error which is THE one being estimated via function approximation of the *state* value function V(s). This makes the definition of the advantage function TABULAR.

Note: The GAE implementation is actually the TRUE GAE (as explained in Sutton, pag. 300) which leverages the fact that the advantage function is TABULAR. In a few experiments on the 6x8 labyrinth it was observed that the TRUE GAE gives better results than GAE. The TRUE GAE requires knowledge of the state value function estimated at the PREVIOUS learning step (t-1) when updating the model parameters for the V(s) NN at time step t. This, however, is approximated as explained in the comments written in LeaTDLambda.learn().

This is a partial commit, so correct execution is not guaranteed.
However, all 73 unit tests pass.

DETAILS:
- td.py: LeaTDLambda
  - New method _updateA_GAE().
  - Method _updateA() now uses the eligibility trace _z_A defined specifically for the advantage function (instead of using _z_Q, the eligibility trace for Q(s,a)).

- value_functions.py, test_optimizers_discretetime.py: changed the definition of the advantage function `A` to TABULAR (based on what is described in the summary above).

- test_optimizers_discretetime.py: the learning rates alpha are now ALWAYS adjusted, because they are used in the learning of the (now always) TABULAR advantage function, as described in the summary above.