[Learners] Changes on the delta(Q) calculation by an improved target value computation
Summary: The expected Q-value of the next state is now computed as the plain average of Q(s,a) over all actions to compute delta(Q) when V(s) and Q(s,a) are NOT TABULAR.
Note that computing the correct pi-weighted Q-value average is currently not possible because the policy is not part of the Learner object responsible for computing such average. However, using the plain average seems to be good enough, and is also backed up by a few articles found on the internet.

This is a partial commit, so correct execution is not guaranteed.
However, all 73 unit tests pass.

DETAILS:
- td.py:
  - LeaTDLambda:
    - _expected_next_Q(s): It now returns V(s) ONLY when both V(s) and Q(s,a) are estimated in the TABULAR context. Otherwise, it returns the PLAIN average of the Q(s,a) values across actions. Reason: Since Q(s,a) and V(s) are estimated with separate models in the non-tabular case, their values may be very different, especially in the average reward context where their estimation is not unique! Actually, I think this problem may also happen in the TABULAR case, but perhaps more mildly...?

    - New method _max_next_Q(s) which returns the maximum Q-value at the next state `s` over all possible next actions. I briefly tried using this method in the calculation of delta(Q), but it seemed to give worse results than using the average of the Q-values.

  - LeaTDLambdaAdaptive:
    - _updateZ(): Fixed the computation of gradient_Q in the AdaptiveLambdaType.HTD case, where V(s) was previously referenced instead of Q(s,a) on the RHS expression that returns grad(Q).