[Learners] Fixed two bugs in the reset of value functions learned by NN models
The ValueFunctionApproxNN class defined in value_functions.py has now the following bugs fixed:
a) Optimizer reset: the torch optimizer was NOT reset by the reset() method.
b) Reset value: the value functions across states (V(s)) and across state and actions (Q(s,a)) were initialized around a positively biased value of 0.1.

Instead, now:
a) the torch optimizer (e.g. Adam) is reset by the reset() method.
b) the value functions are initialized to values distributed around 0.0, thanks to setting the bias of the output neuron to 0.0, instead of 0.1.

The fix done by (a):
- avoids potential divergence of estimated value functions in later replications run as part of a test (as the torch optimizer was previously NOT reset at the start of a new replication!). Of course this was observed when testing the FVAC process on the usual 6x8 labyrinth, o.w. I wouldn't have carried out this fix!
- makes the results of running the same experiment using the same seed always the same, regardless of the running order in which the experiment falls within a sequence of replications run on the same problem.

The fix done by (b) reduces the chances of having a biased estimate of value functions (this is a conjecture, as such behaviour was not observed in experiments).

This is a partial commit, so correct execution is not guaranteed.
However, all 73 unit tests pass.