[Learners] Implemented input dummy neuron for terminal states in NN function approximation
Goal: Give the terminal state the capability of learn its associated values (i.e. V(s) and Q(s,a)) independently of the other states.
This is CRUCIAL in CONTINUING task contexts, where the process is reset to a start state when a terminal state is reached, making the terminal state be adjacent to a start state, something that cannot be inferred by the structure of the environment (e.g. a 2D gridworld whose 2D state is used as input of the neural network model).
If we do not include this neuron, the value of a terminal state is usually NOT properly learned, e.g. in a 2D gridworld (see the tests run in the main section of value_functions.py for an example on a 3x4 2D gridworld).

This is a partial commit, so correct execution is not guaranteed.
However, all 73 unit tests pass.