[Simulator] Anchor action is now RANDOM (crucial for policy learning with NN function approximation)
Goal: Avoid unlearning of the policy which was observed when using always the same a=0 when transitioning from a terminal state to a start state (do not really understand why at this point).
This has particularly a large impact in improving learning of value functions in the NN function approximation approach.

Notes:
1) NO copy of action and advantage value functions is done anymore for terminal states
2) In order to generate the random action, a NEW random number generator is created specifically for that task, in order to avoid modifying trajectories generated by the policy when that action was actually fixed. Note also that, if we were to use the np.random functions to generate the random anchor action, that would still NOT modify the trajectories of the Monte Carlo simulation because the policy actions are chosen using the INDEPENDENT random generator defined by the gym environment! (self.env.np_random --see the np_random() function defined in gym.utils.seeding for details on how this np_random object is created --basically using `np_random = np.random.RandomState()` and then setting the seed `np_random.seed(seed)`).
However, this does NOT apply to the Fleming-Viot simulation because the choice of the FV particle ALSO uses an np.random function, hence its stream is modified by the random anchor action selection.
=> This implies that the estimated value functions should NOT change between before and after the random anchor action implementation!

The unit tests in test_simulators.py were updated, as the change as an impact on the estimated Q values for the terminal state.
Unit tests in test_estimators_discretetime.py involving the Q(s,a) function were updated due to changes in the values of the terminal state.

This is a partial commit, so correct execution is not guaranteed.

Out of the 73 unit tests, 72 pass: the one that fails now is the test on FV with soft killing... but I don't understand why... in fact the state counts change and I don't see why! (because of what is described above regarding the use of a separate random generator with the np_random object created to select the random anchor actions at terminal states.
In any case, the state counts do not change much and the state value function V(s) is almost the same as before, but what changes considerably is the Q(s,a) function (e.g. its mean is 0.11 now and before was 0.14). Still to investigate.