[Simulator] It is now possible to NOT learn Q(s,a) (for speed-up purposes)
To do this, the user should define the dictionary of value functions passed to the learner's constructor with the entry 'Q' set to None.

Took the opportunity to fix the information shown about the update of value functions at every simulation step in debug mode (before this change, the same values were shown before and after each debugged simulation step).

This is a partial commit, so correct execution is not guaranteed.
All 73 unit tests should pass.