[FV] Implemented Keith Ross's approach for the average reward vale used in the advantage function calculation
Summary: The average reward for the MC step of the FV process is now set FIXED to the value observed during the A-estimation exploration, and this follows the algorithm proposed by Keith Ross in his 2021 paper on learning under the average reward criterion, "Average reward reinforcement learning with monotonic policy improvement", where the average reward is first computed over the whole observed trajectory and then used in the advantage value estimation (as opposed to updating the average reward value iteratively as the trajectory is observed).

Goal: Eliminate the volatility sometimes observed in FV learning of the state value function V(s) (with possible eventual divergence), especially for SMALL problems, due to the combination of the following two algorithm implementation aspects:
- the MC step of the FV process (the INI exploration) starts at the external boundary of A which, in small environments like 6x8 gridworld with NO wind, may be close to the state with reward.
- the EPISODIC update of the average reward, that as opposed to an update at every simulation step.

Effect: This change had a crucial impact on improving the stability of FVAC learning in small environments where rewards are NOT rare, such as the usual 6x8 random labyrinth with NO wind. Thanks to the increased stability, the learning rate for the Adam optimizer (lr) and for the Advantage estimation (alpha_initial) can be set a little larger, thus accelerating learning. Ex: to lr=1E-2 and alpha_initial=1.0, instead of 1E-3 and 0.5, respectively (see for instance tests run in the main section of value_functions.py).

Note: I also performed an additional small change in discrete.py: the absorption set A estimated during the initial exploration is now INITIALIZED ALSO when t_learn < 0 (as opposed to only when t_learn = 0). This avoids the comparison of the newly estimated absorption set with a previously stored set A in the FV learner which, when no set A is stored in the learner, makes the number of simulation steps T to be used in the INI exploration (MC step) increase from its initial value of e.g. 500 to its maximum value of e.g. 2000 in one go (as the relative change in size of A is Infinity!). This is particularly important for the unit tests context where t_learn is usually set to t_learn=-1 in order to indicate that the estimation of value functions is NOT part of a policy learning process.

WARNING: All unit tests in test_estimators_discretetime.py on FV now FAIL. I have not updated these tests because I don't have time now, as I am finishing my PhD thesis!
However, test results do NOT change much based on RMSE (so we are most likely in business) (e.g. RMSE = 0.035 now vs. 0.034 before).